Updated practice mode to work with time goal

+++ Lyngua/app/src/main/java/com/example/lyngua/models/goals/Goal.kt
@@ -22,8 +22,8 @@ open class Goal(
     var goalType: Int,
     var numWordsCompleted: Int,
     var totalNumWords: Int,
-    var timeSpent: Int,
-    var totalTime: Int
+    var timeSpent: Double,
+    var totalTime: Double
 
 ) : Parcelable {
 
@@ -39,7 +39,7 @@ open class Goal(
 
         time = myCalendar
         numWordsCompleted = 0
-        timeSpent = 0
+        timeSpent = 0.0
 
     }
